Updated to handle UTF-16-LE LTSpice Rawfiles
LTSpice now outputs rawfiles in UTF-16. All '.decode()' calls in the
document have been changed to '.decode(encoding='utf_16_le',
errors='ignore')'.

The reason we ignore errors is that when we use readline() on the
rawfile, it will return a bytes object that goes up to the '\n'
character. However, since we are using UTF-16-LE, the 0x0A that
represents the newline character is actually one byte before the end of
a line - newlines are encoded as 0x0A00. Python complains that output
from readline did not decode properly because readline() left off the
last byte of the newline.

For this same reason, I've also inserted a line:
raw_file.seek(raw_file.tell() + 1)
after all readline statements. This eats the 0x00 that's leftover from
the previous line so that it isn't sitting at the front of the next
line.

My testing was done from python 3.6.