Adding try: except: clause in the load_step_information
Temperature steps create logs that contain unit information in the format '.step temp=25°C'
In this case, the value cannot be converted to float and generated an exception. 
The previous commit was simply accepting this as a string, but, in order not to break existing code, we try anyway to convert to float, and leave it in text format if it fails.