Allow LTSpice installation in non default location
Currently LTSpice needs to be installed in it's default location. This PR would allow a LTSPICEFOLDER global variable to be used to point to the LTSpice folder in order to allow an installation in a non default location.
Our use case is a cluster, where the home directory has a rather limited quota and commonly things should be installed elsewhere. 
The change should have no effect whatsoever on any existing code/use of the library.