Fix Bug with replace
With new regex the space at the end is included in the match so when we replace it, we also delete the space which causes the new netlist to be invalid. This commit adds a new regex group `'replace'` that does not include the space at the end. We then use that group to get the span of the match. I also removed a set of redundant parathesis around the equal sign matching part of the pattern.